synq: move completion + signature-help probes into semantic::completion
Fifth step of the pass-based refactor. The completion-boundary and
signature-help logic isn't LSP-specific — it's a parse-level probe
over a captured token stream. Moving it into semantic puts it where
other interactive frontends (CLI REPL, embedded editor) can reuse it,
and makes LSP genuinely thin over semantic.

Also fixes a layering violation the previous step left behind:
semantic::analysis was feature-gated on "lsp" because it held
LSP-shaped types (Resolution, DocumentAnalysisData, SymbolIdentity).
Those types are moved back to lsp/analysis_data.rs where they
belong; semantic::analysis now holds only neutral token types
(StoredToken, StoredComment, SemanticToken) and is gated on
"validation" like the rest of semantic.

- new semantic::completion module holds CompletionContext,
  CompletionInfo, completion_info() + helpers, SignatureHelpInfo,
  find_enclosing_call
- completion_info now takes `tokens: &[StoredToken]` directly
  instead of threading through DocumentAnalysisData, decoupling it
  from the LSP bundle shape
- semantic::analysis slimmed to StoredToken + StoredComment +
  SemanticToken only; un-gated from "lsp" to "validation"
- DefinitionLocation / DefinitionResult / ResolvedSymbol /
  Resolution / SymbolIdentity / DocumentAnalysisData relocated
  back to lsp/analysis_data.rs (that's the layer that cares about
  their shape)
- lsp/completion.rs deleted
- LSP SignatureHelpInfo + find_enclosing_call removed from host.rs;
  host now calls semantic::completion::* directly

No behavior change. 244 unit + 34 doctests + 25 integration tests
pass on --features lsp.
